+++ src/jv_unicode.h
@@ -1,14 +1,16 @@
 #ifndef JV_UNICODE_H
 #define JV_UNICODE_H
 
-const char* jvp_utf8_backtrack(const char* start, const char* min, int *missing_bytes);
+#include <stdint.h>
+
+const char* jvp_utf8_backtrack(const char* start, const char* min, intptr_t *missing_bytes);
 const char* jvp_utf8_next(const char* in, const char* end, int* codepoint);
 int jvp_utf8_is_valid(const char* in, const char* end);
 
 int jvp_utf8_decode_length(char startchar);
 
 int jvp_utf8_encode_length(int codepoint);
-int jvp_utf8_encode(int codepoint, char* out);
+intptr_t jvp_utf8_encode(int codepoint, char* out);
 
 int jvp_codepoint_is_whitespace(int c);
 #endif
